+++ database/migrations/2024_12_16_055145_create_compras_table.php
@@ -15,20 +15,20 @@ class CreateComprasTable extends Migration
     public function up()
     {
         Schema::create('compra', function (Blueprint $table) {
-            $table->bigIncrements('id_compra'); 
-            $table->unsignedBigInteger('id_proveedor'); 
-            $table->dateTime('fecha_compra', 3)->default(DB::raw('CURRENT_TIMESTAMP(3)')); 
-            $table->double('total_compra', 10, 2); 
-            $table->double('descuento_compra', 10, 2)->default(0.0); 
-            $table->boolean('factura_compra')->default(false); 
-            $table->timestamps(); 
+            $table->bigIncrements('id_compra');
+            $table->unsignedBigInteger('id_proveedor');
+            $table->dateTime('fecha_compra', 3)->default(DB::raw('CURRENT_TIMESTAMP(3)'));
+            $table->decimal('total_compra', 10, 2);
+            $table->decimal('descuento_compra', 10, 2)->default(0.0);
+            $table->boolean('factura_compra')->default(false);
+            $table->timestamps();
 
             // Llave foránea
             $table->foreign('id_proveedor')
-                  ->references('id_proveedor')
-                  ->on('proveedor')
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
+                ->references('id_proveedor')
+                ->on('proveedor')
+                ->onDelete('cascade')
+                ->onUpdate('cascade');
         });
     }
