+++ database/migrations/2024_12_16_055403_create_detalle_compras_table.php
@@ -14,26 +14,26 @@ class CreateDetalleComprasTable extends Migration
     public function up()
     {
         Schema::create('detalle_compra', function (Blueprint $table) {
-            $table->bigIncrements('id_detalle_compra'); 
-            $table->unsignedBigInteger('id_compra'); 
-            $table->unsignedBigInteger('id_producto'); 
-            $table->string('descripcion', 191)->nullable(); 
-            $table->integer('cantidad_compra'); 
-            $table->double('subtotal_compra', 10, 2); 
-            $table->timestamps(); 
+            $table->bigIncrements('id_detalle_compra');
+            $table->unsignedBigInteger('id_compra');
+            $table->unsignedBigInteger('id_producto');
+            $table->string('descripcion', 191)->nullable();
+            $table->decimal('cantidad_compra', 10, 2);
+            $table->decimal('subtotal_compra', 10, 2);
+            $table->timestamps();
 
             // Llaves foráneas
             $table->foreign('id_compra')
-                  ->references('id_compra')
-                  ->on('compra')
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
+                ->references('id_compra')
+                ->on('compra')
+                ->onDelete('cascade')
+                ->onUpdate('cascade');
 
             $table->foreign('id_producto')
-                  ->references('id_producto')
-                  ->on('producto')
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
+                ->references('id_producto')
+                ->on('producto')
+                ->onDelete('cascade')
+                ->onUpdate('cascade');
         });
     }
